Move console under registrar
