+++ internal/kernel/cli.go
@@ -3,9 +3,9 @@ package kernel
 import (
 	"os"
 
-	"github.com/daison12006013/gorvel/console"
 	"github.com/daison12006013/gorvel/pkg/env"
 	"github.com/daison12006013/gorvel/pkg/errors"
+	"github.com/daison12006013/gorvel/registrar"
 	cli "github.com/urfave/cli/v2"
 )
 
@@ -15,7 +15,7 @@ func ConsoleApplication() {
 	app := &cli.App{
 		Name:     "craft",
 		Usage:    "A crafting console command tool for garvel!",
-		Commands: *console.Commands,
+		Commands: *registrar.Commands,
 	}
 
 	err := app.Run(os.Args)
