+++ resources/views/users/_table.go.html
@@ -73,7 +73,7 @@
             <tbody class="bg-white divide-y divide-gray-200">
               {{ if lt $.data.Paginate.Total 1 }}
                 <tr>
-                  <td colspan="3" class="px-6 py-4 whitespace-nowrap text-gray-800 text-sm text-center">
+                  <td colspan="5" class="px-6 py-4 whitespace-nowrap text-gray-800 text-sm text-center">
                   No records found.
                   </td>
                 </tr>
